// src/types/ScanState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ScanState = { enabled: boolean, corners: [[number, number], [number, number], [number, number], [number, number]], edges: [[number, number], [number, number], [number, number], [number, number]], };